Reorganized project structure to store database files only in /data
